Allow for single-level module_path when packing (#371)
* Allow for single-level module_path when packing

Prior to 0.7.1, a module_path of `./lib` would result in a package
containing files like this (using `test/app8` as an example):

    lib/app8.node

The switch to tar 3 in 0.7.1 causes the same package to end up like
this:

    package.json
    app8.cc
    binding.gyp
    index.js
    README.md
    lib/app8.node

which causes a problem during remote binary installation:

    node-pre-gyp ERR! Pre-built binaries not installable for PACKAGE@0.0.1 and node@8.3.0 (node-v57 ABI, glibc) (falling back to source compile with node-gyp)
    node-pre-gyp ERR! Hit error ENOTDIR: Cannot cd into '..dirs/package/lib'

This commit fixes this problem and allows for single-level module_path
values.

* Compare apples to apples

But don't compare Linux to Apple.

* Use a platform-specific dynamic object extension

This should work everywhere except AIX.

* Fix Win32 file expectations

This is based on what showed up during an Appveyor run. Someone with
more Win32 module building experience should verify that these are
expected and correct.